Assert on meck/file/code functions; purge test mod
It is necessary to assert on some return values of the
aforementioned modules because without that we don't know
whether they succeeded or not.

Also Erlang VM allows to load only 2 versions of a particular
module. Because of that after the 3rd case there was an error
stating that code:load_file/1 failed. With code:purge/1 in the
end_per_testcase this is fixed.