Copy code to Doker context while building image instead of cloning
Previously amoc source code was retrieved in Dockerfile with cloning
code from github. It did not allow to trigger automated builds
for other branch than master, because it was always cloning master.

To avoid it, we can copy code into build image. To make it possible
it is necessary to move Dockerfile into project root directory. It
allows to use `COPY . /amoc_build`.

There is also new file .dockerignore which is similiar to .gitignore.
It allows to specify files which should not be copied into build context
while building docker image.