Make it possible to add users in batches
There is new `amoc_controller:add_batches/2` function that allows to add
users in batches according to some strategy. The function takes the
scenario module and the number of batches as parameters.

The strategy of adding users should be returned by `next_user_batch/2`
callback implemented in the scenario module. The callback is optional.
The batch index and the number of users added in the previous batch are
parmeters that are passed to the callback function.  The strategy is a
list of {Node, NumOfUsers, Interarrival}.

Batches are added every batch interval specified by `add_batch_interval`
application environment variable, which is 5 minutes by default.

Adding batches can be sheduled via HTTP API by specifying `batches` key
in a body request to `scenarios/$SCENARIO` endpoint.