Pool user supervisors
If users go up and down too fast, amoc_users_sup becomes a bottleneck as it
can’t keep up with the requests. Improve the performance by creating a pool of
supervisors whose size is proportional to the number of schedulers, and ensure
distribution of users among them by for example hashing on the user_id.

We do this by creating a pool of supervisors and having the whole API through
`amoc_users_sup_sup` choose supervisors according to requests.