Implement send_and_wait as a blocking gen_server:call/3
Notice that a normal send would spawn a local process, send a pid across
distribution, have the local process receive an atom across distribution, and
have the `Msg' be delivered to the requester locally.

That is 1 pid and 1 atom across distribution and one local payload, plus one
spawn that copies the payload.

Doing a call instead will only send back and forth the payload, and already
solve monitoring problems

This is specially optimal when the payload is a literal, like an atom.