Allow any term to be a coordinator name
This comes from when the name was the name of a registered process, and
hence the Erlang registry required it be an atom, but now it is an
opaque term stored in a persistent term, hence allowing the flexibility
for it to be any term allows us to create coordinators dynamically.

One example of when this might be desired is when coordinating members
of a groupchat: when the groupchat is created, a coordinator is created
with the name of the groupchat, and the members will add themselves to
this coordinator. This way, they can coordinate setting presences and
messages within the namespace of that room alone.