Refactor gradual config logic away from the controller
Not only this simplifies how the gradual config is triggered, but it
also makes the gradual config transparent to the controller: the
controller just applies a new rate every so much time. This not just
makes the controller much simpler, but also, makes it possible to give
it rate changes that are not necessarily only linear, but that could
have more interesting shapes.